requirements.txt
python-dotenv
PyPDF2
PyMuPDF
groq
ratelimit
anthropic